refactor: refact the impl structure for different os
