+++ monoio/src/fs/permissions.rs
@@ -81,6 +81,6 @@ impl Debug for Permissions {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         f.debug_struct("Permissions")
             .field("readonly", &self.readonly())
-            .finish()
+            .finish_non_exhaustive()
     }
 }
